Added Guessing Game Functionality
Implemented functionality in the Python script to allow users to guess the randomly generated number. The program tracks the number of guesses made by the user and provides feedback based on whether the guess is correct or not. If the user enters a non-numeric value, an error message is displayed, prompting them to enter a valid number.